requirements.txt
discord.py[voice]
python-dotenv
youtube-dl
requests
uvloop